Plant-Pathology-2020-Kaggle/code/requirements.txt
absl-py==0.9.0
appnope==0.1.0
astor==0.8.1
backcall==0.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
cycler==0.10.0
Cython==0.29.15
decorator==4.4.1
future==0.18.2
gast==0.3.3
glob2==0.7
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
idna==2.9
imageio==2.8.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
joblib==0.14.1
jupyter-client==5.3.4
jupyter-core==4.6.1
kaggle==1.5.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.1.3
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.14.6
olefile==0.46
pandas==1.0.1
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.5.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
PySocks==1.7.1
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2019.3
pyzmq==18.1.1
requests==2.23.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
tb-nightly==1.14.0a20190301
tensorflow==2.0.0a0
termcolor==1.1.0
text-unidecode==1.2
tf-estimator-nightly==1.14.0.dev2019030115
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
tqdm==4.43.0
traitlets==4.3.3
Unidecode==1.1.1
urllib3==1.24.3
wcwidth==0.1.8
Werkzeug==1.0.0
